Add extra LineChartViewBase abstraction for Line Chart & Area Chart
