+++ appinventor/appengine/src/com/google/appinventor/client/editor/simple/components/MockAreaChartView.java
@@ -1,6 +1,6 @@
 package com.google.appinventor.client.editor.simple.components;
 
-public class MockAreaChartView extends MockLineChartView {
+public class MockAreaChartView extends MockLineChartViewBase {
     public MockAreaChartView() {
         super();
     }
