+++ appinventor/appengine/src/com/google/appinventor/client/editor/simple/components/MockLineChartView.java
@@ -2,17 +2,12 @@ package com.google.appinventor.client.editor.simple.components;
 
 import org.pepstock.charba.client.ScatterChart;
 
-public class MockLineChartView extends MockChartView<ScatterChart> {
+public class MockLineChartView extends MockLineChartViewBase {
     /**
      * Creates a new MockLineChartView object instance.
      */
     public MockLineChartView() {
-        // A ScatterChart widget is used to allow
-        // arbitrary insertions of (x, y) points and
-        // automatic generation of the X axis labels
-        // (as well as automatic scaling)
-        chartWidget = new ScatterChart();
-        initializeDefaultSettings();
+        super();
     }
 
     @Override
